python3.9/cookiecutter-aws-sam-cdk-hello-python-datadog/{{cookiecutter.project_name}}/requirements.txt
aws-cdk-lib
constructs
